Fix: Optimise deprecation stage (#50)
* add file exists check for phpstan.neon file and move output into general if statement

* fix: add missing slash in file exists check

+++ lib/stages/deprecation.sh
@@ -1,16 +1,19 @@
 #!/usr/bin/env bash
 
 _stage_deprecation() {
-    printf "Checking for deprecations.\n\n"
-
-    local project_location
-    local composer_bin_dir
-
-    project_location=$(get_project_location)
-    composer_bin_dir=$(get_composer_bin_directory)
-
     if ${DRUPAL_TESTING_TEST_DEPRECATION}; then
-        cp phpstan.neon "${DRUPAL_TESTING_DRUPAL_INSTALLATION_DIRECTORY}"
+        printf "Checking for deprecations.\n\n"
+
+        local project_location
+        local composer_bin_dir
+
+        project_location=$(get_project_location)
+        composer_bin_dir=$(get_composer_bin_directory)
+
+        if [ ! -f "${DRUPAL_TESTING_DRUPAL_INSTALLATION_DIRECTORY}/phpstan.neon" ]; then
+            cp phpstan.neon "${DRUPAL_TESTING_DRUPAL_INSTALLATION_DIRECTORY}"
+        fi
+
         cd "${DRUPAL_TESTING_DRUPAL_INSTALLATION_DIRECTORY}" || exit
         "${composer_bin_dir}/phpstan" analyse --memory-limit "${PHPSTAN_MEMORY_LIMIT}" "${project_location}"
         cd - || exit
